+++ README.md
@@ -1,8 +1,3 @@
 # sc2-angular
 SteemConnect2 with AngularJS
 
-# Installation
-
-    npm install
-    npm start
-    # application should be available at: http://localhost:3000/
