--- package.json
@@ -1,19 +1,0 @@
-{
-  "name": "sc2-angular",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "postinstall": "gulp build",
-    "start": "node ./bin/www"
-  },
-  "dependencies": {
-    "body-parser": "~1.17.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.3",
-    "express": "~4.15.2",
-    "gulp": "^3.9.1",
-    "morgan": "~1.8.1",
-    "sc2-sdk": "0.0.1",
-    "serve-favicon": "~2.4.2"
-  }
-}
